fix(member): 🐛 增加了判空并修改返回值

+++ src/main/java/org/netunion/manager/controller/MemberController.java
@@ -11,7 +11,7 @@ import java.util.List;
  * 成员页面控制器
  *
  * @author David Wang
- * @version 1.2
+ * @version 1.3
  * @date 2020-07-07
  */
 @RestController
@@ -26,7 +26,8 @@ public class MemberController {
     // 获取指定成员信息 JSON 格式
     @GetMapping("/member/{id}")
     public Member getMember(@PathVariable("id") int id) {
-        return memberMapper.getById(id);
+        if (memberMapper.getById(id) == null) return null;
+        else return memberMapper.getById(id);
     }
     //通过 POST 方式添加成员
     @PostMapping(value = "/member/add", produces = "application/json;charset=UTF-8")
@@ -37,19 +38,38 @@ public class MemberController {
     //通过 DELETE 方式删除成员
     @DeleteMapping(value = "/member/delete/{id}")
     public String deleteMember(@PathVariable("id") int id) {
-        memberMapper.delete(id);
-        return "删除成员成功";
+        Member member = memberMapper.getById(id);
+        if (member != null){
+            memberMapper.delete(id);
+            return member.toString();
+        }
+        else {
+            return "EMPTY ID";
+        }
     }
     //通过 POST 方式更新成员银行卡号
     @PostMapping(value = "/member/update/bank")
     public String updateMemberBankNum(@RequestParam int id, @RequestParam String bankNum) {
-        memberMapper.updateBankNumById(id, bankNum);
-        return "更新成员银行卡号成功";
+        Member member = memberMapper.getById(id);
+        if (member == null) {
+            return "EMPTY ID";
+        }
+        else {
+            memberMapper.updateBankNumById(id, bankNum);
+            return member.toString();
+        }
     }
     //通过 POST 方式更新成员手机号码
     @PostMapping(value = "/member/update/phone")
     public String updateMemberPhoneNum(@RequestParam int id, @RequestParam String phoneNum) {
-        memberMapper.updatePhoneNumById(id, phoneNum);
-        return "更新成员手机号码成功";
+        Member member = memberMapper.getById(id);
+        //判空
+        if (member == null) {
+            return "EMPTY ID";
+        }
+        else {
+            memberMapper.updatePhoneNumById(id, phoneNum);
+            return member.toString();
+        }
     }
 }
